Remove particle weight (p.wt) override after getNextParticle() call
Rather than replacing the weight of a particle by the source weight, the code should now multiply the source weight by the particle weight after a new particle is generated.  Thus, sources which return new particles with weights which are not unity (such as egs_collimated_source) should now work in egs_brachy.