Remove technical manual PDF (#63)
* Remove reference to PDF of technical manual

PDF of the technical manual is no longer generated. Users should use the HTML version (online or in the docs/ folder).

* Fix #57 Delete empty egs_brachy_technical_manual.pdf

Generation of the PDF of the technical manual has been broken for a while and only produces a blank PDF. The PDF is removed and users should use the HTML version instead.